# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trackventory API",
        "description": "API documentation for Trackventory - Inventory Tracking System",
        "contact": {
            "name": "Trackventory Support",
            "email": "support@trackventory.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/request-statuses": {
            "get": {
                "tags": [
                    "Request Statuses"
                ],
                "summary": "Get list of request statuses",
                "description": "Retrieve all request statuses",
                "operationId": "getRequestStatusesList",
                "responses": {
                    "200": {
                        "description": "Request statuses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request statuses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RequestStatus"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Request Statuses"
                ],
                "summary": "Create a new request status",
                "description": "Store a newly created request status",
                "operationId": "storeRequestStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestStatusStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request status created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request status created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RequestStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/request-statuses/{id}": {
            "get": {
                "tags": [
                    "Request Statuses"
                ],
                "summary": "Get request status information",
                "description": "Retrieve a specific request status by ID",
                "operationId": "getRequestStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request status retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RequestStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request status not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Request Statuses"
                ],
                "summary": "Update request status",
                "description": "Update the specified request status",
                "operationId": "updateRequestStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestStatusUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RequestStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request status not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Request Statuses"
                ],
                "summary": "Delete request status",
                "description": "Remove the specified request status",
                "operationId": "deleteRequestStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Request status deleted successfully"
                    },
                    "404": {
                        "description": "Request status not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestStatus": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status_name": {
                        "type": "string",
                        "example": "Approved"
                    },
                    "badge_color": {
                        "type": "string",
                        "example": "#28A745"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "name": [
                                "The name field is required."
                            ],
                            "email": [
                                "The email field must be a valid email address."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "RequestStatusStoreRequest": {
                "required": [
                    "status_name",
                    "badge_color"
                ],
                "properties": {
                    "status_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Approved"
                    },
                    "badge_color": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "green"
                    }
                },
                "type": "object"
            },
            "RequestStatusUpdateRequest": {
                "properties": {
                    "status_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Approved"
                    },
                    "badge_color": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "green"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}